requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
djangorestframework==3.15.2
kombu==5.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.1
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
